refactor: split config.rs (3959 lines) into focused submodules
Replace the monolithic commands/config.rs with a module directory:

- config/mod.rs   — shared helpers + 7 core/gateway commands + re-exports
- config/ai.rs    — 8 AI provider commands
- config/mcp.rs   — 11 MCP server commands + private mcp helpers
- config/channels.rs — 8 channel commands (Telegram, Feishu, etc.)
- config/agents.rs   — 9 agent commands + 6 agent-related structs
- config/settings.rs — 23 settings commands + 8 config structs

Submodules access shared helpers via super::. Rust resolves pub mod config
to config/mod.rs automatically, so main.rs requires no changes.
